Exclusively use int instead of float to avoid OpenGL issues

+++ gui/app.h
@@ -25,9 +25,9 @@ public:
         virtual void onOpenGLActivate(int width, int height) {}
         virtual void onDraw() {}
         virtual void onResize(int width, int height) {}
-        virtual void onScroll(float dx, float dy) {}
-        virtual void onLeftMouseDown(float mouse_x, float mouse_y, app::ModifierKey modifiers) {}
-        virtual void onLeftMouseDrag(float mouse_x, float mouse_y, app::ModifierKey modifiers) {}
+        virtual void onScroll(int dx, int dy) {}
+        virtual void onLeftMouseDown(int mouse_x, int mouse_y, app::ModifierKey modifiers) {}
+        virtual void onLeftMouseDrag(int mouse_x, int mouse_y, app::ModifierKey modifiers) {}
         virtual void onKeyDown(app::Key key, app::ModifierKey modifiers) {}
         virtual void onClose() {}
         virtual void onDarkModeToggle() {}
